requirements.txt
Cheetah==2.4.4
configobj==5.0.6
Django==1.10.5
django-cors-headers==2.1.0
Jinja2==2.9.5
jsonpatch==1.15
jsonpointer==1.10
Markdown==2.6.8
MarkupSafe==0.23
mysqlclient==1.3.7
oauth==1.0.1
olefile==0.44
raven==6.7.0
# pip install pillow requires a lot of dev packages, thus we install pillow directly by 'apk add py2-pillow'
# current py2-pillow version is 4.3.0
Pillow>=4.1.0
prettytable==0.7.2
python-memcached==1.58
PyYAML==3.12
requests==2.13.0
six==1.10.0
celery[redis]==3.1.25
django-celery==3.2.1
python-alipay-sdk==1.7.0
gunicorn==19.7.1
redis==2.10.6
sqlparse==0.2.4
grequests==0.3.0
aliyun-python-sdk-core==2.13.4
